Split SQL INSERT query to avoid reaching MySQL's max_allowed_packet
max_allowed_packet default value is 64MB for MySQL and 16MB for MariaDB

When this limit is reached, the connection is closed and we end up with
"MySQL server has gone away" error

This limit can be reached when updating a resource that have a lot of
text in its values. This can happen easily with modules like
ExtractText.

This patch splits an SQL INSERT query so that it runs one query per
value instead of one query per resource.
While it's not perfect and the limit still can be reached by a single
value, it should make it less easy to be reached.