Fix Doctrine "Unable to find entity identifier"
Doctrine tracks in the EM/UoW even the references you create, so
references that are going to be persisted need to come from the EM
that's going to do the persisting.

This fix changes CurrentReferenceMetadata to optionally take the entity
manager as an argument, so it can use the "private" EM used by the job.